Refactor remove method in AppointmentController to use async/await

+++ src/modules/appointments/services/remove.Appointment.service.ts
@@ -1,10 +1,25 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, Injectable } from '@nestjs/common';
 import AppointmentRepository from '../typeorm/repositories/AppointmentRepository';
 
 @Injectable()
 export class RemoveAppointmentService {
   constructor(private readonly appointmentRepository: AppointmentRepository) {}
-  remove(id: string) {
-    return this.appointmentRepository.delete(id);
+  async remove(id: string): Promise<void> {
+    const appointment = await this.appointmentRepository.findOne(id);
+
+    console.log(appointment);
+
+    if (!appointment) {
+      throw new HttpException('Consulta não encontrada', 404);
+    }
+
+    if (appointment.status !== 'scheduled') {
+      throw new HttpException(
+        'Apenas consultas com status "agendado" podem ser removidos',
+        400,
+      );
+    }
+
+    await this.appointmentRepository.delete(id);
   }
 }
